nsq cleanup
